requirements.txt
django
djangorestframework-jwt
djangorestframework
drf_yasg
mysqlclient
mysql-connector-python
django-silk
django-cors-headers